Refactor cache, auth, storage dependencies/file structure
