+++ src/app/admin/uploads/[uploadPath]/page.tsx
@@ -3,7 +3,7 @@ import AdminChildPage from '@/components/AdminChildPage';
 import { PATH_ADMIN, PATH_ADMIN_UPLOADS } from '@/site/paths';
 import { extractExifDataFromBlobPath } from '@/photo/server';
 import { redirect } from 'next/navigation';
-import { getUniqueTagsCached } from '@/cache';
+import { getUniqueTagsCached } from '@/photo/cache';
 
 interface Params {
   params: { uploadPath: string }
