+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import { getPhotosCached, getPhotosCountCached } from '@/cache';
+import { getPhotosCached, getPhotosCountCached } from '@/photo/cache';
 import AnimateItems from '@/components/AnimateItems';
 import MorePhotos from '@/photo/MorePhotos';
 import SiteGrid from '@/components/SiteGrid';
@@ -11,7 +11,7 @@ import {
 } from '@/site/pagination';
 import { pathForRoot } from '@/site/paths';
 import { Metadata } from 'next';
-import { MAX_PHOTOS_TO_SHOW_OG } from '@/photo/image-response';
+import { MAX_PHOTOS_TO_SHOW_OG } from '@/image-response';
 
 export const runtime = 'edge';
