+++ src/app/tag/[tag]/[photoId]/layout.tsx
@@ -10,7 +10,7 @@ import {
   absolutePathForPhotoImage,
 } from '@/site/paths';
 import PhotoDetailPage from '@/photo/PhotoDetailPage';
-import { getPhotoCached } from '@/cache';
+import { getPhotoCached } from '@/photo/cache';
 import { getPhotosTagDataCached } from '@/tag/data';
 import { ReactNode } from 'react';
