+++ src/auth/cache.ts
@@ -1,0 +1,4 @@
+import { cache } from 'react';
+import { auth } from '@/auth';
+
+export const authCached = cache(auth);
