+++ src/components/OGTile.tsx
@@ -5,7 +5,7 @@ import { clsx } from 'clsx/lite';
 import Link from 'next/link';
 import { BiError } from 'react-icons/bi';
 import Spinner from '@/components/Spinner';
-import { IMAGE_OG_DIMENSION } from '../photo/image-response';
+import { IMAGE_OG_DIMENSION } from '../image-response';
 
 export type OGLoadingState = 'unloaded' | 'loading' | 'loaded' | 'failed';
