+++ src/image-response/HomeImageResponse.tsx
@@ -1,5 +1,5 @@
 import { SITE_DOMAIN_OR_TITLE } from '@/site/config';
-import { Photo } from '..';
+import { Photo } from '../photo';
 import ImageCaption from './components/ImageCaption';
 import ImageContainer from './components/ImageContainer';
 import ImagePhotoGrid from './components/ImagePhotoGrid';
