+++ src/image-response/components/ImageCaption.tsx
