+++ src/image-response/components/ImageContainer.tsx
