+++ src/image-response/components/ImagePhotoGrid.tsx
