+++ src/image-response/index.ts
